fix: loads env for local build
