+++ build.gradle
@@ -26,7 +26,7 @@ dependencies {
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
 
     implementation 'io.projectreactor.kotlin:reactor-kotlin-extensions'
-    implementation 'io.github.cdimascio:dotenv-java:2.2.0'
+    implementation 'io.github.cdimascio:dotenv-java:3.0.0'
     implementation 'com.okta.spring:okta-spring-boot-starter:3.0.6'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
